refactor(user): change from users to user
