+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/appetis.iml" filepath="$PROJECT_DIR$/.idea/appetis.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/workly.iml" filepath="$PROJECT_DIR$/.idea/workly.iml" />
     </modules>
   </component>
 </project>
