+++ packages/api/package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "nodemon": "^2.0.7",
-    "pm2": "^4.5.6"
+    "pm2": "^4.5.6",
+    "sequelize-cli": "^6.2.0"
   }
 }
