+++ packages/api/server/controllers/user.http
@@ -1,0 +1,15 @@
+POST http://localhost:3000/api/users
+Content-Type: application/json
+
+{"email": "test@test.com", "password": "test"}
+
+###
+GET http://localhost:3000/api/users
+Accept: application/json
+Authorization: Bareer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpYXQiOjE2MjU0NDk5NTcsImV4cCI6MTYyNTQ1MTc1NywiaXNzIjoiYXBwZXRpcyJ9.i1yzX3u7-hWvYSc-0qf8YU1pRTaXRJDiP0iiVN_6Am4
+###
+GET http://localhost:3000/api/users/1
+Accept: application/json
+Authorization: Bareer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3RAdGVzdC5jb20iLCJpYXQiOjE2MjU0NDk5NTcsImV4cCI6MTYyNTQ1MTc1NywiaXNzIjoiYXBwZXRpcyJ9.i1yzX3u7-hWvYSc-0qf8YU1pRTaXRJDiP0iiVN_6Am4
+
+###
